# expression: PlusInt check null iff overflow