Signed-off-by: TennyZhuang <zty0826@gmail.com>

<!-- Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

Issue Number: close #xxx <!-- REMOVE this line if no issue to close -->

Problem Summary:

Some simple arithmetic function like PlusInt, check nulls every run, and the check is more expensive then the compute.
The bitmaps have been already merged, so the check is unnecessary, we can compute it anyway while there is no overflow , and we don't care the result.

### What is changed and how it works?

Avoid some branch in `PlusInt`

I've done a trivial benchmark:


before

```
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-VecBuiltinFunc
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-VecBuiltinFunc-12                    	  251072	      4881 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-NonVecBuiltinFunc
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-NonVecBuiltinFunc-12                 	   32326	     35392 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-VecBuiltinFunc#01
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-VecBuiltinFunc#01-12                 	  505320	      2382 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-NonVecBuiltinFunc#01
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-NonVecBuiltinFunc#01-12              	   41050	     28024 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-VecBuiltinFunc#02
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-VecBuiltinFunc#02-12                 	  446617	      2730 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-NonVecBuiltinFunc#02
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-NonVecBuiltinFunc#02-12              	   42690	     34795 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-VecBuiltinFunc#03
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-VecBuiltinFunc#03-12                 	  375538	      2868 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-NonVecBuiltinFunc#03
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-NonVecBuiltinFunc#03-12              	   43404	     26849 ns/op	       0 B/op	       0 allocs/op
```

After:

```
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-VecBuiltinFunc
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-VecBuiltinFunc-12                     	  537535	      2169 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-NonVecBuiltinFunc
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-NonVecBuiltinFunc-12                  	   37428	     31334 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-VecBuiltinFunc#01
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-VecBuiltinFunc#01-12                  	  806791	      1464 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-NonVecBuiltinFunc#01
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-NonVecBuiltinFunc#01-12               	   42577	     26617 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-VecBuiltinFunc#02
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-VecBuiltinFunc#02-12                  	  487568	      2404 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-NonVecBuiltinFunc#02
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-NonVecBuiltinFunc#02-12               	   44211	     27624 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-VecBuiltinFunc#03
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-VecBuiltinFunc#03-12                  	  442723	      2425 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-NonVecBuiltinFunc#03
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusIntSig-NonVecBuiltinFunc#03-12               	   41484	     28918 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticModRealSig-VecBuiltinFunc
```

10% - 100% performance improvement in different cases.

### Related changes

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test

Side effects

### Release note <!-- bugfixes or new feature need a release note -->

- No release note